Move resource package inside calendar
