+++ mobile/src/test/java/com/alexstyl/specialdates/events/namedays/calendar/NamedayJSONParserTest.java
@@ -1,7 +1,7 @@
 package com.alexstyl.specialdates.events.namedays.calendar;
 
 import com.alexstyl.specialdates.events.namedays.NamedayLocale;
-import com.alexstyl.specialdates.events.namedays.resource.NamedayJSONParser;
+import com.alexstyl.specialdates.events.namedays.calendar.resource.NamedayJSONParser;
 
 import org.junit.Test;
